[FIX] survey: ensure same color in community and enterprise while taking surveys
Use $primary to always get same color in COM and ENT. The goal of using color
variable for styling is to get always the same color, no matter if we are in
community or on enterprise.

Before this commit, *$o-brand-primary* was used in survey stylesheet but this
variable is not matching the same colors in COM and in ENT.

After this commit, *$primary* is used to render the same color no matter the
environment (COM or ENT).

Triggered by task ID 1972640 (live sessions on survey)
PR #44205