[IMP] event: improve sale computation on event
Purpose
=======

In this commit we set the total sales price of an event to be the converted
sum of order lines, tax excluded. Currency is now a related on the company to
avoid de-synchronizing events and companies.

Fields and computation comes from 95abffde52770e6fc0b7cb1153e70c98fb1fb44f .

Task 2170831
PR #43173