[FIX] web_editor: fix color palette custom color behaviors
Selecting a css color equals to a theme color via the custom color
was not selecting the theme color.

Opening and closing the custom color without picking a color was
unselecting the selected color.

Now the custom color picker has the selected color as default even if it
does not appear in the palette as selected. This was appearing in the
parallax snippet as no custom color is displayed.

closes odoo/odoo#44986

X-original-commit: b04e711344a4381b6f482c0735f49cc6f6a9ff75
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>