[FIX] stock: _search_picking_for_assignation and immediate transfer
When a move is created through a push rule, the system will first try to
find a picking to put it into before creating a new one. It is possible
the move is placed into an immediate picking where the reservation
fields are hidden and the initial demand is updated according to the
qty_done with [0].

As this behaviour is unwanted, do not merge move into immediate
transfers (which should be short-lived anyway).

[0] 7a0c30f7277e24a9dd1e2821bfc8a26147bd198c

closes odoo/odoo#44086

X-original-commit: fa9efd28e6e113cd9993a29bca070ddee83630c0
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>