[FIX] web: fix boolean fields display in pivot view
Since formatBoolean returns a jQuery object, boolean fields were
displayed as [Object object] in pivot view.

so fix the issue of checkbox in the pivot view by not calling
fromatValue for boolean instead just set the checkbox element in
in template itself and added test case for checkbox in pivot view.

closes odoo/odoo#46445

Taskid: 2172258
Closes: #44299
X-original-commit: 4f0a7611941bd144f7419a4c8ef048054bbdbe39
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>