[FIX] stock: readonly fields independant of 'show detailled'
The 'Show Detailled Operation' setting on picking type makes the
entire operation tab readonly if set to 'True'.
This could be confusing as the description and date expected fields are
readonly depending on this setting.

This commit makes sure the stock move fields keep their readonly
attribute (or not) independently of the show detailled setting.

Task : 2180347