[FIX] mrp: wrong MO finished date
Usecase to reproduce:
- Create a product with 10 days manufacture lead time
- Create a procurement in order to fullfill a need the 01/15

The MO start date is the 01/05 that's the expected values
The MO finished date is the 01/05, 1 h later than the start date

It happens due to commit afef7fbb322a262855ba5ebb16d173d231b4445f
This commit modified the wrong start date in order to take the
procurement date - manufacturing lead time. However it also
modified the finished date to 'the start date + 1h'. That returns
the problem to the opposite direction. Now the finished moves have
a wrong planning.

Fix it by using the production + company lead days when it's set
instead of 1h.

opw-2181962

X-original-commit: 6a5ce8de91d3ab2cb7f73127a6ee1c11578ccef9